Add files via upload
Space Forces is a videogame based on math, that takes the user across 3 different levels in space, following the challenges of an astronaut both in third and first person. The objective of the game is to increase the mental math skills of the user by practicing basic operations while having fun in an exciting game!

The game idea, images, edits, design of the levels, and all the coding in C# were done 100% by me.